[v1.1] typeset -f <functname>: generate script using sh_deparse()
Problem reproducer:
- Locate or write a script with a function definition
- Source the script with '.' or 'source'
- Print the function definition with 'typeset -f'; looks good
- Edit the file to add or remove characters before or in the
  function definition, or rename or remove the file
- Print the function definition with 'typeset -f' again; the output
  is corrupted!

Cause: ksh remembers the offset[*1] and length[*2] of the function
definition in the source file, and prints function definitions by
simply dumping that amount of bytes from that offset in the source
file. This will obviously break if the file is edited or (re)moved,
so that approach is fundamentally flawed and needs to be replaced.

All other shells that are capable of printing function definitions
(bash, mksh, yash, zsh) regenerate the function definition from the
parse tree. ksh does include code for a deparser (in deparse.c) but
it is disused (so it was removed in the 1.0 branch). This commit
changes 'typeset -f' to use it, as well as 'typeset -T' when
printing type discipline functions.

It also applies lots of tweaks and fixes to the deparser. More
fixes might be needed. So, it's time to test this.

[*1] functloc in struct functnod --> hoffset in struct Ufunction
[*2] repurposed functline in struct functnod --> nvsize in
     struct Namval

src/cmd/ksh93/sh/deparse.c:
- Too many tweaks to mention for more correct script generation and
  better formatting; this is the result of intermittent work over
  about a year so I don't quite remember everything.
- sh_deparse(): Add argument for initial indentation level.
- p_keyword(): Convert the BEGIN/MIDDLE/END argument into a bitmask
  and add a NOTAB bit flag to avoid printing a tab after a keyword.

src/cmd/ksh93/sh/string.c: sh_fmtq():
- For assignments, avoid quoting variable names containing a dot,
  e.g. .namespace.foo=bar should not become '.namespace.foo=bar'.
- Avoid quoting the variable name in additive assignments, e.g.
  foo+=bar\ baz should become foo+='bar baz', not 'foo+=bar baz'.

src/cmd/ksh93/include/shnodes.h,
src/cmd/ksh93/sh/parse.c:
- funct(): Do not save function location, offset and length.

src/cmd/ksh93/bltins/typeset.c: print_namval():
- For 'typeset -f functionname', instead of retrieving the function
  definition from the saved position in its source file, call
  sh_deparse() to generate script from the parse tree. Start
  indentation at level 0.

src/cmd/ksh93/bltins/typeset.c: sh_outtype():
- For type discipline functions output by 'typeset -T typename',
  instead of retrieving the function definition from the saved
  position in its source file, call sh_deparse() to generate script
  from the parse tree. Start indentation at level 1.