test -e: do not return true for nonexistent pseudodevices
The -e option of test/[/[[ returns true for pseudodevices that are
supported by the shell, but do not exist on the file system.

Astonishingly, this change was an intentional addition in ksh 93u-:

| 10-08-24 +[[ -e /dev/xxx/ ]] can be used to check whether special
|          files of those names are handled by the shell.

For instance, '[[ -e /dev/tcp/ ]]', as wel as 'test -e /dev/tcp/',
returns true, though it does not exist in the file system.

Another thing that ksh supports internally is /dev/fd/* file
descriptor devices. These are also supported by most OS kernels,
but not by, e.g., AIX or Haiku. On those systems we get:

  $ test -e /dev/fd/0 && cat /dev/fd/0
  cat: /dev/fd/0: No such file or directory

...which should *never* happen. Even on systems with /dev/fd (e.g.
Linux, *BSD, macOS), it breaks: 'test -e' returns true for closed
file descriptors, which do not have a /dev/fd/* file system entry.

So this misfeature is now reverted as it breaks the basic idea of
the test -e option and is contrary to documented and standardised
behaviour.

Thanks to @JohnoKing for finding and reporting this problem.
Resolves: https://github.com/ksh93/ksh/issues/548