Fix fallback string field splitting regression (re: 6c73c8c3)
@JohnoKing reports:
> Reproducer (this should print 'foo', not an error):
>
>   $ unset NOTSET
>   $ ${NOTSET:-echo -e foo}
>   /usr/bin/ksh: echo -e: not found

The problem is that field splitting is not happening before the
last argument that contains a -, ^ or !.

  $ printf '[%s]\n' ${UNSET:-a b c xxx-xxx d e f}
  [a b c xxx-xxx]
  [d]
  [e]
  [f]

Analysis: the problem is caused by these lines:

  src/cmd/ksh93/sh/macro.c
  863:	if(c)
  864:		sfwrite(stkp,first,c);
  865:	first = fcseek(c);

That writes the text from first up to the current character onto
the stack without processing field splitting. This is done
unconditionally by the new S_BRAOP code, which explains the bug.

src/cmd/ksh93/sh/macro.c: copyto():
- As field splitting is not done in quoted contexts, fix the bug by
  simply doing nothing and continuing to the next character unless
  the context is a bracket expression that is quoted with either
  single or double quotes.

Resolves: https://github.com/ksh93/ksh/issues/553