rm export of attributes via A__z (re: 00d4396, 7954855, ff4e912c)
A rather astonishing and little known feature of ksh93 is that not
only the values, but also the type attributes (if any) are exported
when you export a variable. It uses a special environment variabele
called A__z to communicate the attributes between ksh instances. No
other shell does this, and virtually no one expects it. In my view,
this is a misfeature that causes more harm than good.

Export/import of variable attributes for newly invoked scripts
rarely (if ever) seems to be used intentionally, but does sometimes
cause real problems for newly invoked scripts in child processes
that use the same variable names as variables exported by the
shell invoking them. For example, if the parent shell does
	integer SOMEVAR
	export SOMEVAR
then a script invoked as a command by that shell will probably not
expect SOMEVAR to be of an integer type without it having typeset
it as such. If it uses a variable by that name, it may break.

Though I've never seen one, there are probably a few scripts that
depend on this, which is why it was kept in 1.0 for compatibility
reasons (and disabled in POSIX mode) but is now removed as of 1.1.

Resolves: https://github.com/ksh93/ksh/issues/566