cleanup: libast: remove preroot support
preroot is a kernel facility that existed (as far as I can
determine) only in the Amdahl UTS variant of UNIX for mainframes,
which has been discontinued (last 4.5 release was late 1990s or
early 2000s). Acccording to man/preroot.3:

> preroot is a kernel supported per-process two level viewpath.
> All pathnames rooted at / are first searched for in the process
> preroot directory and then in the system root directory. Setting
> the process preroot is a privileged operation controlled by the
> /etc/preroot(1) command.

The very few mentions of /etc/preroot on Google are a 1987 Usenet
post and a 2000 forum post. High time to remove this dead code.

https://en.wikipedia.org/wiki/Amdahl_UTS
https://usenet.trashworldnews.com/?thread=508157
http://computer-programming-forum.com/53-perl/b19c3497d9f099aa.htm