break/continue: remove odd 'label not implemented' error message
When giving an invalid number argument to 'break' or 'continue',
ksh gives an odd error message:

$ while :; do break 1badnum; done
-ksh: break: 1badnum: label not implemented

This has been unchanged since at least 1995. Perhaps there was an
idea to implement labels to jump to in ksh93, which was abandoned.

This commit removes that message and uses e_badnumber instead, so
the error message is more informative and more like other shells:

$ while :; do break 1badnum; done
-ksh: break: 1badnum: bad number