Delete < 1995 junk code from tminit to fix `printf %(%Z)T` on Cygwin (#615)
The TZ variable is commonly set to values such as 'America/Los_Angeles',
which makes it a terrible choice when attempting to obtain the timezone
name with printf '%(%Z)T'. In some scenarios the output returned by
printf %T is completely bogus because it assumes $TZ will be set to a
short string (e.g., "MST"); in practice this is rarely the case. On
Cygwin, attempting to use %Z to get the timezone does the following for
PST ($TZ is set to "America/Los_Angeles"):
  $ date +%Z
  PDT
  $ printf '%(%Z)T\n' now
  ica
The 'ica' was extracted from the end of 'America' in $TZ. For %Z, that
result is complete nonsense (the result from date(1) is 'PDT', which
printf %T should match). This bug currently causes two regression
tests in builtins.sh to fail on Cygwin.

Minor note: This bug is only relevant for $TZ, not $TZNAME. While
$TZNAME is fairly uncommon, it should have the correct set of names when
set (e.g., $TZNAME will be "PST,PDT" for Pacific Time, which the current
code handles just fine).

src/lib/libast/tm/tminit.c:
- Fix the above problem by removing the bogus code that obtains 'ica'
  and other nonsense strings from $TZ.