package, shtests: add csh workaround
The 1995 ksh93 distribution had a trick to allow invoking the build
script directly from csh without adding an unportable #! path. This
commit adds a tweaked version of that:

min_posix=/if/this/is/csh/ignore/the/error/message || exec sh $0:q $argv:q

This works because it can be parsed by both sh and csh. On (t)csh,
the sh assignment is interpreted as a command name, which of course
doesn't exist, so throws an error:

min_posix=/if/this/is/csh/ignore/the/error/message: Command not found.

The error cannot be suppressed, at least not in a way that will not
make the sh parser error out, so this is the best we can do. At
least (t)csh users can now build and test with the standard
commands without having to prefix 'sh'.

Also, in bin/shtests, the code to escape from an antique Bourne or
broken POSIX shell (see 20f8557c) is copied over from bin/package.