edit.c: --nomultiline: adjust for prompt length when redrawing (#633)
In both line editors, starting the shell or resizing the window
when multiline is set to off causes graphical errors for the first
line of text not reached by a subsequent carriage return or Ctrl +
C. The correct position for scrolling is three spaces away from the
window edge; when the error occurs, the text instead continues for
the length of the prompt (e_prompt) before scrolling begins. For
example, if $COLUMNS is 90 and e_prompt is 20 characters long, text
scrolls after 87 characters rather than 67. This results in partial
multiline output (depending on how long the prompt is) and, on some
operating systems, the repetition of the command line.

The cause is in the function ed_read in edit.c. When the shell
starts or the window is resized, the new size-associated variables
are determined through ed_read rather than ed_setup. Because the
code in ed_read only accounts for prompt length when multiline is
on, the effective window size exceeds its correct value by the
number of characters in the prompt (e_plen).

src/cmd/ksh93/edit/edit.c: ed_read():
- Adjust for the length of the prompt when determining the new
  window size when multiline is set to off, provided that the
  result will be greater than zero.

Resolves: https://github.com/ksh93/ksh/issues/632