build system: mamake: allow 'prev' to declare a prerequisite
In the Mamfiles, dependencies on prerequisites (files that are
expected to be present, and that outdate the target if their
timestamp gets updated, but are not generated by the current make
action -- for example, *.c source files) were generally declared
using empty make...done blocks. This threw me at first; I thought
the prev command could be used for that. But a prev command without
a prior make...done block just creates an empty dummy rule and is
effectively ignored.

There is an opportunity here to make the Mamfiles more concise.
This commit makes a prev without a prior make...done equivalent to
an empty make...done, allowing the use of prev to define a new rule
declaring a dependency on a simple prerequisite. In this case, any
attributes given to prev are added to the new rule and not ignored.

The change is backward compatible by default. Setting MAMAKE_STRICT
now enables backward incompatible strict checking: for prev
commands declaring prerequisites, a nonexistent prerequisite is an
error (unless the virtual or dontcare attribute is given), and for
prev commands referencing previous make...done rules, attributes
are an error ('superfluous attributes') instead of being ingored.

src/cmd/INIT/mamake.c:
- strict(): Added. Returns true if the MAMAKE_STRICT mamake
  variable is set.
- make(): case KEY('p','r','e','v'):
  * Handle the case where a prior make...done rule doesn't exist:
    . Don't simply call rule() which adds an empty rule if not
      found; instead, process the attributes with a call to
      attributes() and, unless the 'virtual' attribute was given,
      bind the rule to the prerequisite file.
    . If strict(), and the keepgoing state (-k option) is not set,
      make a nonexistent prerequisite an error, else a warning.
  * For the case where a prior rule does exist, attributes to prev
    are ignored because the pre-existing attributes are used; if
    strict(), now throw an error if attributes were given to a prev
    command that references a pre-existing rule.

**/Mamfile:
- Set MAMAKE_STRICT variable.
- Convert all empty make...done blocks to prev commands, combining
  the attributes that were given to make and done.
- Remove the ignored attributes from all old prev commands.
- libast/Mamfile: Fix erroneous dependencies of the virtual
  joint.conflim.h and joint.lc.h targets. They used prev to declare
  a dependency on conflim.h and lc.h before these were finished
  generating, which was ineffective. (This was evidently a bug in
  nmake, which originally generated this code.)