Update doc after backporting ${x/#/text} fix to 1.0 (re: ceae1e44)
With the 1.0.5 release around the corner, I am reconsidering the
decision not to have this fix on the 1.0 branch.

On 1.0 and legacy ksh, ${x/#/y} is a no-op; it's equivalent to $x.
There is no use case for that, so there is no good reason why any
legacy script would use it – *except*, of course, if the author
expected it to work. In which case, fixing this in ksh means fixing
that script. Fixing it will therefore likely do more good than
harm.

The existing behaviour may have been technically correct (an empty
glob pattern matches nothing even if anchored), but that is a very
obscure kind of technically correct. Fixing it will align ksh
93u+m/1.0.5 with reasonable user expectations (particularly since
this has been working as expected for many years on mksh, zsh and
bash).

Between that, and the fact that there is no use case for the
original behaviour, fixing it seems like the right thing to do –
particularly as the 1.0 branch has already seen fixes for clear
design flaws (such as the octal leading zero mess, or matching '.'
and '..' in globbing) that are actually more radical than this one.