tests/printf.sh: Fix regress tests for 32-bit (re: 1a9d8ada)
GNU date(1) will not provide a date for 20 or 40 years into the
past or future on 32-bit systems, because this exceeds the limits
of seconds encoded in a 32-bit signed integer (which will last
until 19 Jan 2038). libast does not have a problem with this, but
the GNU date limitation does cause spurious test failures.

src/cmd/ksh93/tests/printf.sh:
- do_test(): If GNU date(1) fails to provide a value, print a
  warning and skip the test.
- Remove a bunch of duplicate GNU date(1) comparison tests.