fun: Tidy up the _cd/dirs/mcd/popd/pushd mess
This is supposed to be a coherent installation, but the functions
in 'fun' did not play well together and were not well organised.
The main issues were:
- pushd and popd had duplicate definitions.
- To get 'mcd' you had to autoload 'dirs'.
- There were two different versions of 'dirs': one in dirs, one in
  pushd and popd. The first version prints the stack with numbers.

This commit reorganises the functions so that there is one main
function per file, as one would expect when autoloading functions
via FPATH, and fixes them to play well together so that everything
just works when the default fun directory is in FPATH (as is now
done by 'bin/package use', see 6ffab556).

It also reformats the indenting of these functions to match the
standard style of this code base, and applies various tweaks,
including quoting for safety and printing error messages to
standard error.

We keep only one version of 'dirs' (the one with numbers, as this
comes in handy when using the 'cd' stack wrapper, e.g. 'cd 2'). So
'pushd' and 'popd' will now show those too, as they call 'dirs'.

The '_cd' wrapper function is renamed to 'cd' and no longer relies
on an a 'cd' alias; this workaround was only needed on ksh version
that are now ancient.

Note that the 'cd' autoloadable wrapper function is not loaded
automatically even if it is on FPATH, because 'cd' is a builtin.
But, on init, dirs/mcd/popd/pushd now run 'autoload cd' so that
'cd' will play well with the directory stack as soon as it is used.

'autocd' is amended to use 'cd' instead of 'command cd' so that it
will now use the directory stack's 'cd' wrapper if loaded, thus
it will now also play well with the directory stack.

'man' is amended so it still shows the self-doc for a built-in even
if a function overrides it; this will avoid a generic external
manual page being shown for 'cd' when 'cd' is overridden.

Finally, this commit adds missing AT&T AST copyright headers to the
originally AT&T-supplied functions.