posix: disable special handling of /dev/fd/n in test/[/[[
On systems without /dev/fd (e.g. Haiku, AIX, HP-UX) or broken
/dev/fd (FreeBSD 13.1):

  $ ksh -c 'test -e /dev/fd/9 9</dev/null && echo positive'
  positive

This should only output 'positive' on systems where /dev/fd/9 is
actually created in the file system when file descriptor 9 is
opened. In other words, it *should* be an effective test for
whether /dev/fd exists and is functional on the current OS.

Unfortunately, ksh is special-casing /dev/fd here (and so is bash).
Perhaps they do this because both ksh and bash can use, e.g.,
> /dev/fd/9 in redirections even if /dev/fd/9 doesn't physically
exist.

But that doesn't mean you can use it for everything: it won't work
with external commands. test -e should always physically test
whether a file exists or not, anything else is just broken and
violates POSIX.

Unfortunately there may well be scripts that depend on this flawed
design to test if they can use /dev/fd in redirections, or to test
whether a file descriptor is open. A quick google already shows
some evidence of that on stackexchange. So a fix must unfortunately
be limited to the POSIX mode.

src/cmd/ksh93/bltins/test.c: sh_access():
- When POSIX mode is on, do not call sh_isdevfd() to decide whether
  to handle strings matching /dev/fd/* via sh_ioaccess() which
  tests for access to an open file descriptor.

A regression test for this is not really feasible since, on systems
with a functional /dev/fd, the special-casing is redundant and the
results should be identical without it.