sh.1: clarify 'nameref var=$1' documentation for functions
Only when the positional parameter is assigned directly to the
nameref as part of the declaration command can it access one
function's local variable from within another. Due to static
scoping, this is not normally possible.

This special-casing is handled in setall() in bltins/typeset.c,
lines 972-976 as of this commit.

Resolves: https://github.com/ksh93/ksh/issues/593