Fix error in special bltin disabling .sh.tilde disc (re: 3c2b0dcf)
Since sh_exit() in fault.c now sets sh.tilde_block to 0, an error
in a special built-in will set that to 0. Then the tilde expansion
completes and sh.tilde_block is decreased in tilde_expand2().
Meaning it gets decreased after being set to 0, so it gets
permanently blocked due to its value -1.

The simplest fix is to explicitly set sh.tilde_block to 1 and 0
instead of increasing and decreasing it.