Make the code -Wall compliant on gcc 12.2 & 13.2 and clang 14.0.0
@kloczek writes:
> Just summary stats using gcc 13.x
>
> [tkloczko@pers-jacek SPECS]$ rpmbuild -ba ksh.spec  2>&1 |
>			grep -- \\[-W | sed 's/.*\[//; s/\]//' |
>			sort | uniq -c | sort -nr
>    2034 -Wunknown-pragmas
>    1226 -Wparentheses
>     283 -Wmissing-braces
>     179 -Wunused-value
>      25 -Wchar-subscripts
>      19 -Wlto-type-mismatch
>       9 -Wunused-variable
>       9 -Wunused-const-variable=
>       2 -Waddress
>       1 -Wuse-after-free
>       1 -Wunused-function
>       1 -Wunused-but-set-variable
>       1 -Wsequence-point
>       1 -Wdangling-pointer=

Cause: build system uses -Wall to turn on ALL the warnings.
Some users want a clean compile with -Wall, so here it is.

-Wunknown-pragmas was fixed by adding preprocessor directives that
hid them from inapplicable compilers. (AFAIK the standard says
unknown pragmas are ignored, so this was pretty obnoxious.)

-Wparentheses and -Wmissing-braces complain about perfectly valid
C90 syntax, and I am not going to change it. Instead, these
warnings have been disabled in src/lib/libast/features/common.

On my own gcc tests I also had a lot of -Wmaybe-unitialized
warnings. In most cases it was obvious from reading the code that
the variables will not be used uninitialised, so those warnings are
likely false positives. But a few of those did look like actual
problems, and the overhead of variable initialisers is negligible,
so this fixes them all.

The rest of the warnings were fixed with appropriate code tweaks,
too many to list here.

Resolves: https://github.com/ksh93/ksh/issues/699