autocd: fix for dir names equal to reserved words (re: 39c5e6a3)
It was not possible to autocd to a directory whose name is
identical to that of a reserved word in ksh -- 'select', for
example. It should be possible by quoting/escaping at least one
character of the word/directory name, but:

  $ autocd
  autocd activated -- type a directory path to cd, ~- to cd back
  type 'autocd -d' to deactivate
  $ mkdir select
  $ \select
  ksh: select: not found

The problem was in the use of 'command -v' to detect whether the
directory name is a command; it will return reserved words. The fix
is to use 'whence -q' instead; the -q option not only suppresses
output but also disables the search for reserved words.

After the fix:

  ...
  $ \select
  cd -- select