build: ksh93: some iffe and SHOPT.sh streamlining
In ksh93/Mamfile there are four iffe invocations that don't run
a script but run one 'def' command directly from the iffe command
line. The 'def' command causes 8 probes to be run (per iffe --man):

    def name
          Equivalent to cmd,dat,hdr,key,lib,mth,sys,typ name.

This was quite inefficient as only one or two of them were used.

The process is streamlined by eliminating the individual iffe
invocations and adding only the probes whose results are actually
used to FEATURE/options (the test for sys/acct.h) and
FEATURE/externs (the rest).

The execargs.h probe and the code depending on it (in main.c) are
removed as this is only relevant for the PDP-11 (RIP, 1970-1997).

This commit also has a bugfix for SHOPT.sh processing in
ksh93/Mamfile. It was not effective to set SHOPT_GLOBCASEDET to 0
on systems where case-insensitive file systems are detected (e.g.
macOS). The fix is to include FEATURE/options in shopt.h after, not
before, the SHOPT.sh-based defines.